+++ content_service/src/main/java/domain/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package domain.define;
+package domain.entity;
 
 
 import jakarta.persistence.Column;
