+++ content_service/src/main/java/domain/entity/user/repository/UserContentRepository.java
@@ -1,7 +1,7 @@
-package domain.define.user.repository;
+package domain.entity.user.repository;
 
-import domain.define.user.UserContent;
+import domain.entity.user.UserContent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserContentRepository extends JpaRepository<UserContent, Long> ,UserContentRepositoryCustom{
+public interface UserContentRepository extends JpaRepository<UserContent, Long> {
 }
